🚚 Mod: 가게 리뷰 조회 폴더 이동 (#5)
