+++ src/index.ts
@@ -3,10 +3,9 @@ import express, { Express, Request, Response } from "express";
 import cors from "cors";
 import { handleUserSignUp } from "./modules/users/controllers/user.controller.js";
 import { handleAddStore } from "./modules/stores/controllers/store.controller.js";
-import { handleAddReview, handleListMyReviews } from "./modules/reviews/controllers/review.controller.js";
+import { handleAddReview, handleListMyReviews, handleListStoreReviews } from "./modules/reviews/controllers/review.controller.js";
 import { handleAddMission } from "./modules/missions/controllers/mission.controller.js";
 import { handleChallengeMission } from "./modules/missions/controllers/challenge.controller.js";
-import { handleListStoreReviews } from "./modules/stores/controllers/storereview.controller.js";
 
 // 1. 환경 변수 설정
 dotenv.config();
