+++ src/modules/reviews/repositories/review.repository.ts
@@ -29,6 +29,30 @@ export const getReview = async (reviewId: number) => {
   return review;
 };
 
+// 가게 리뷰 목록 조회
+export const getAllStoreReviews = async (store_id: number, cursor: number) => {
+  const reviews = await prisma.review.findMany({
+    select: {
+      id: true,
+      body: true,
+      store: true,
+      member: true,
+    },
+    where: {
+      store_id,
+      id: {
+        gt: cursor,
+      },
+    },
+    orderBy: {
+      id: "asc",
+    },
+    take: 5,
+  });
+
+  return reviews;
+};
+
 // 내가 작성한 리뷰 조회
 export const getMyReviews = async (memberId: number, cursor: number) => {
   const reviews = await prisma.review.findMany({
