--- src/modules/stores/dtos/storeReview.dto.ts
@@ -1,22 +1,0 @@
-export interface ReviewItem {
-  id: number;
-  body: string | null;
-}
-
-export const responseFromReviews = (reviews: ReviewItem[]): ReviewListResponse => {
-  const lastReview = reviews[reviews.length - 1];
-
-  return {
-    data: reviews,
-    pagination: {
-      cursor: lastReview ? lastReview.id : null,
-    },
-  };
-};
-
-export interface ReviewListResponse {
-  data: ReviewItem[];
-  pagination: {
-    cursor: number | null;
-  };
-}
